ER-1367 Renamed test project so it gets executed in azure pipeline +semver: patch
