+++ src/Jobs/Recruit.Vacancies.Jobs.UnitTests/DomainEvents/Handlers/Provider/ProviderBlockedDomainEventHandlerTests/GivenEmployerOwnedLiveVacancies.cs
@@ -9,12 +9,12 @@ using Moq;
 using Newtonsoft.Json;
 using Xunit;
 
-namespace Esfa.Recruit.Vacancies.Jobs.Tests.DomainEvents.Handlers.Provider.ProviderBlockedDomainEventHandlerTests
+namespace Esfa.Recruit.Vacancies.Jobs.UnitTests.DomainEvents.Handlers.Provider.ProviderBlockedDomainEventHandlerTests
 {
-    public class GivenEmployersWithPermissionsOnly : ProviderBlockedDomainEventHandlerTestBase
+    public class GivenEmployerOwnedLiveVacancies : ProviderBlockedDomainEventHandlerTestBase
     {
         [Fact]
-        public async Task ShouldNotifyEmployersAboutRevokedPermission()
+        public async Task ShouldNotifyEmployersAboutLiveVacancy()
         {
             var employerAccount1 = "EmployerAccount1";
             var employerAccount2 = "EmployerAccount2";
@@ -29,38 +29,35 @@ namespace Esfa.Recruit.Vacancies.Jobs.Tests.DomainEvents.Handlers.Provider.Provi
                 QaVacancyUser = new VacancyUser()
             };
 
-            var vacancies = GetVacancies(OwnerType.Employer, employerAccount1, 1, VacancyStatus.Live);
-            var info = GetEmptyProviderProfile(employerAccount1, employerAccount2, employerAccount3);
-            var sut = GetSut(info, vacancies);
+            var vacancies = GetVacancies(OwnerType.Employer, employerAccount1, 1, VacancyStatus.Live)
+                .Concat(GetVacancies(OwnerType.Employer, employerAccount2, 1, VacancyStatus.Live))
+                .Concat(GetVacancies(OwnerType.Employer, employerAccount3, 1, VacancyStatus.Draft));
+
+            var sut = GetSut(GetEmptyProviderProfile(employerAccount1), vacancies);
             var payload = JsonConvert.SerializeObject(data);
             await sut.HandleAsync(payload);
-
+            
             MockCommunicationQueueService.Verify(q => q.AddMessageAsync(It.Is<CommunicationRequest>(c => 
-                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForPermissionOnly
+                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForLiveVacancies 
                 )), Times.Exactly(2));
 
             MockCommunicationQueueService.Verify(q => q.AddMessageAsync(It.Is<CommunicationRequest>(c => 
-                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForPermissionOnly &&
+                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForLiveVacancies &&
+                c.ParticipantsResolverName == CommunicationConstants.ParticipantResolverNames.EmployerParticipantsResolverName && 
+                c.TemplateProviderName == CommunicationConstants.ServiceName && 
+                c.DataItems.Count == 0 && 
+                c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Employer && e.EntityId.ToString() == employerAccount1) &&
+                c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Provider && (long)e.EntityId == ukprn)
+                )));
+
+            MockCommunicationQueueService.Verify(q => q.AddMessageAsync(It.Is<CommunicationRequest>(c => 
+                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForLiveVacancies &&
                 c.ParticipantsResolverName == CommunicationConstants.ParticipantResolverNames.EmployerParticipantsResolverName && 
                 c.TemplateProviderName == CommunicationConstants.ServiceName && 
                 c.DataItems.Count == 0 && 
                 c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Employer && e.EntityId.ToString() == employerAccount2) &&
                 c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Provider && (long)e.EntityId == ukprn)
                 )));
-
-            MockCommunicationQueueService.Verify(q => q.AddMessageAsync(It.Is<CommunicationRequest>(c => 
-                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForPermissionOnly &&
-                c.ParticipantsResolverName == CommunicationConstants.ParticipantResolverNames.EmployerParticipantsResolverName && 
-                c.TemplateProviderName == CommunicationConstants.ServiceName && 
-                c.DataItems.Count == 0 && 
-                c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Employer && e.EntityId.ToString() == employerAccount3) &&
-                c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Provider && (long)e.EntityId == ukprn)
-                )));
-
-            MockCommunicationQueueService.Verify(q => q.AddMessageAsync(It.Is<CommunicationRequest>(c => 
-                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForPermissionOnly &&
-                c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Employer && e.EntityId.ToString() == employerAccount1)
-                )), Times.Never);
         }
     }
 }
