+++ src/Jobs/Recruit.Vacancies.Jobs.UnitTests/DomainEvents/Handlers/Provider/ProviderBlockedDomainEventHandlerTests/GivenEmployersWithPermissionsOnly.cs
@@ -9,12 +9,12 @@ using Moq;
 using Newtonsoft.Json;
 using Xunit;
 
-namespace Esfa.Recruit.Vacancies.Jobs.Tests.DomainEvents.Handlers.Provider.ProviderBlockedDomainEventHandlerTests
+namespace Esfa.Recruit.Vacancies.Jobs.UnitTests.DomainEvents.Handlers.Provider.ProviderBlockedDomainEventHandlerTests
 {
-    public class GivenProviderOwnedVacancies : ProviderBlockedDomainEventHandlerTestBase
+    public class GivenEmployersWithPermissionsOnly : ProviderBlockedDomainEventHandlerTestBase
     {
         [Fact]
-        public async Task MustNotifyEmployersWithTransferCount()
+        public async Task ShouldNotifyEmployersAboutRevokedPermission()
         {
             var employerAccount1 = "EmployerAccount1";
             var employerAccount2 = "EmployerAccount2";
@@ -29,36 +29,38 @@ namespace Esfa.Recruit.Vacancies.Jobs.Tests.DomainEvents.Handlers.Provider.Provi
                 QaVacancyUser = new VacancyUser()
             };
 
+            var vacancies = GetVacancies(OwnerType.Employer, employerAccount1, 1, VacancyStatus.Live);
+            var info = GetEmptyProviderProfile(employerAccount1, employerAccount2, employerAccount3);
+            var sut = GetSut(info, vacancies);
             var payload = JsonConvert.SerializeObject(data);
-            var vacancies = GetVacancies(OwnerType.Provider, employerAccount1, 2, VacancyStatus.Live)
-                .Concat(GetVacancies(OwnerType.Provider, employerAccount2, 1, VacancyStatus.Draft))
-                .Concat(GetVacancies(OwnerType.Employer, employerAccount3, 1, VacancyStatus.Draft));
-                
-            var sut = GetSut(GetEmptyProviderProfile(), vacancies);
             await sut.HandleAsync(payload);
 
             MockCommunicationQueueService.Verify(q => q.AddMessageAsync(It.Is<CommunicationRequest>(c => 
-                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForTransferredVacancies
+                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForPermissionOnly
                 )), Times.Exactly(2));
 
-
             MockCommunicationQueueService.Verify(q => q.AddMessageAsync(It.Is<CommunicationRequest>(c => 
-                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForTransferredVacancies &&
+                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForPermissionOnly &&
                 c.ParticipantsResolverName == CommunicationConstants.ParticipantResolverNames.EmployerParticipantsResolverName && 
                 c.TemplateProviderName == CommunicationConstants.ServiceName && 
-                c.DataItems.Any(d => d.Key == CommunicationConstants.DataItemKeys.Employer.VacanciesTransferredCount && d.Value == "2") && 
-                c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Employer && e.EntityId.ToString() == employerAccount1) &&
-                c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Provider && (long)e.EntityId == ukprn)
-                )));
-
-            MockCommunicationQueueService.Verify(q => q.AddMessageAsync(It.Is<CommunicationRequest>(c => 
-                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForTransferredVacancies &&
-                c.ParticipantsResolverName == CommunicationConstants.ParticipantResolverNames.EmployerParticipantsResolverName && 
-                c.TemplateProviderName == CommunicationConstants.ServiceName && 
-                c.DataItems.Any(d => d.Key == CommunicationConstants.DataItemKeys.Employer.VacanciesTransferredCount && d.Value == "1") && 
+                c.DataItems.Count == 0 && 
                 c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Employer && e.EntityId.ToString() == employerAccount2) &&
                 c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Provider && (long)e.EntityId == ukprn)
                 )));
+
+            MockCommunicationQueueService.Verify(q => q.AddMessageAsync(It.Is<CommunicationRequest>(c => 
+                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForPermissionOnly &&
+                c.ParticipantsResolverName == CommunicationConstants.ParticipantResolverNames.EmployerParticipantsResolverName && 
+                c.TemplateProviderName == CommunicationConstants.ServiceName && 
+                c.DataItems.Count == 0 && 
+                c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Employer && e.EntityId.ToString() == employerAccount3) &&
+                c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Provider && (long)e.EntityId == ukprn)
+                )));
+
+            MockCommunicationQueueService.Verify(q => q.AddMessageAsync(It.Is<CommunicationRequest>(c => 
+                c.RequestType == CommunicationConstants.RequestType.ProviderBlockedEmployerNotificationForPermissionOnly &&
+                c.Entities.Any(e => e.EntityType == CommunicationConstants.EntityTypes.Employer && e.EntityId.ToString() == employerAccount1)
+                )), Times.Never);
         }
     }
 }
